requirements.txt
flask==3.0.0
line-bot-sdk==3.7.0
python-dotenv